Moving data_directory fixture to a best location
