+++ ibis/backends/tests/conftest.py
@@ -1,5 +1,6 @@
 import importlib
 import os
+from pathlib import Path
 from typing import List
 
 import pandas as pd
@@ -158,6 +159,25 @@ def pytest_pyfunc_call(pyfuncitem):
 
 
 pytestmark = pytest.mark.backend
+
+
+@pytest.fixture(scope='session')
+def data_directory() -> Path:
+    """
+    Fixture that returns the test data directory.
+
+    Returns
+    -------
+    Path
+        Test data directory
+    """
+    root = Path(__file__).absolute().parent.parent.parent.parent
+
+    default = root / 'ci' / 'ibis-testing-data'
+    datadir = os.environ.get('IBIS_TEST_DATA_DIRECTORY', default)
+    datadir = Path(datadir)
+
+    return datadir
 
 
 @pytest.fixture(params=_get_backends_to_test(), scope='session')
